Profile Page separation
